# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: RydeWorks Dispatch Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Trips ──────────────────────────────────────────────────────────
...
  /admin/vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vehicles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehiclesListResponse"

    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create vehicle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleResponse"
  /admin/vehicles/{id}:
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleRequest"
      responses:
        "200":
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleResponse"

  # ── Users ──────────────────────────────────────────────────────────
  /admin/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users/drivers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"

    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete/deactivate user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin / Org ──────────────────────────────────────────────────────────
  /admin/org:
    get:
      operationId: getOrganization
      tags: [admin]
      summary: Get organization settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"

    patch:
      operationId: updateOrganization
      tags: [admin]
      summary: Update organization settings (branding, fare zones, etc.)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"

  # ── Reports ──────────────────────────────────────────────────────────
  /admin/reports/trips:
    get:
      operationId: getTripsReport
      tags: [reports]
      summary: Get trips report
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Trips report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"
  /admin/reports/drivers:
    get:
      operationId: getDriversReport
      tags: [reports]
      summary: Get drivers report
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Drivers report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResponse"

  # ── Payments ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
      required: [success, error]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [success, token, user]
    User:
      type: object
      properties:
        _id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: string
        organization:
          type: string
        isActive:
          type: boolean
      required: [_id, firstName, lastName, email, roles]
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required: [success, user]
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required: [success, users]
    CreateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            type: string
      required: [firstName, lastName, email, password, roles]
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    Vehicle:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: number
        licensePlate:
          type: string
        color:
          type: string
        capacity:
          type: number
        status:
          type: string
        baseLocation:
          type: object
          additionalProperties: true
      required: [_id, name]
    VehicleResponse:
      type: object
      properties:
        success:
          type: boolean
        vehicle:
          $ref: "#/components/schemas/Vehicle"
      required: [success, vehicle]
    VehiclesListResponse:
      type: object
      properties:
        success:
          type: boolean
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
      required: [success, vehicles]
    CreateVehicleRequest:
      type: object
      properties:
        name:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: number
        licensePlate:
          type: string
        color:
          type: string
        capacity:
          type: number
      required: [name]
    UpdateVehicleRequest:
      type: object
      properties:
        name:
          type: string
        make:
          type: string
        model:
          type: string
        licensePlate:
          type: string
        color:
          type: string
        capacity:
          type: number
        status:
          type: string
    Organization:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        email:
          type: string
        phone:
          type: string
        logo:
          type: string
        primaryColor:
          type: string
        accentColor:
          type: string
        appName:
          type: string
        plan:
          type: string
        homeBases:
          type: array
          items:
            type: object
            additionalProperties: true
        fareZones:
          type: array
          items:
            type: object
            additionalProperties: true
        selfPayConfig:
          type: object
          additionalProperties: true
        settings:
          type: object
          additionalProperties: true
      required: [_id, name, slug]
    OrganizationResponse:
      type: object
      properties:
        success:
          type: boolean
        org:
          $ref: "#/components/schemas/Organization"
      required: [success, org]
...
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        logo:
          type: string
        primaryColor:
          type: string
        accentColor:
          type: string
        appName:
          type: string
        homeBases:
          type: array
          items:
            type: object
            additionalProperties: true
        fareZones:
          type: array
          items:
            type: object
            additionalProperties: true
        selfPayConfig:
          type: object
          additionalProperties: true
        settings:
          type: object
          additionalProperties: true
...
    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        summary:
          type: object
          additionalProperties: true
      required: [success, data]
...