# lib/api-zod/src/generated/types/reportResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RydeWorks Dispatch Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { ReportResponseDataItem } from "./reportResponseDataItem";
import type { ReportResponseSummary } from "./reportResponseSummary";

export interface ReportResponse {
  success: boolean;
  data: ReportResponseDataItem[];
  summary?: ReportResponseSummary;
}