# lib/api-zod/src/generated/types/usersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RydeWorks Dispatch Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { User } from "./user";

export interface UsersListResponse {
  success: boolean;
  users: User[];
}